backend/requirements.txt
# Framework Web Django
django==5.2.4
djangorestframework==3.16.0
django-cors-headers==4.7.0

# Traitement des données
pandas==2.3.1
numpy==2.3.1

# Upload de fichiers
pillow==11.3.0
python-multipart==0.0.20

# Génération de rapports PDF
reportlab==4.0.7
fpdf2==2.7.6

# Utilitaires
python-dotenv==1.0.0

# Tests (optionnel)
pytest==7.4.3
pytest-django==4.8.0
